save-demo
### What's done:
* Added new spring service - save-demo
* Moved everything connected with demos  into external demo spring service
* Disabled security for SaveDemo

(#1415)
