+++ save-demo/src/main/kotlin/com/saveourtool/save/demo/service/DiktatDemoService.kt
@@ -1,4 +1,4 @@
-package com.saveourtool.save.sandbox.service
+package com.saveourtool.save.demo.service
 
 import com.saveourtool.save.demo.diktat.DiktatDemoAdditionalParams
 import com.saveourtool.save.demo.diktat.DiktatDemoMode
