+++ save-demo/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+server.port=5401
+#management.endpoints.web.exposure.include=health,info,prometheus
+spring.codec.max-in-memory-size=100MB
